wids-datathon-2020/requirements.txt
# local package
-e .

# external requirements
Click==7.0
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
ipykernel

# project requirements
pandas==1.0.1
pyarrow
scikit-learn==0.22.2
catboost
seaborn
shap
matplotlib
plotly
chart-studio
missingno
pytest
streamlit